requirements.txt
dj-database-url==0.5.0
django==2.2.4
djangorestframework==3.11.0
gunicorn==19.9.0
mysqlclient==1.4.4
pytz==2019.2
sqlparse==0.3.0
whitenoise==4.1.3
